Initial upload of my shandroid bot for Robocode competition
Shandroid -- the one that ruled them all at the IndyJug Robocode competition for two years straight.  No cheating here, this was an open source mega-bot based on Diamond.  Substantial changes were put in place, and some interesting experiments with bullet shields, but in the end, none of that really helped.